pyproject.toml
[tool.poetry]
name = "simbirsoft-internship"
version = "0.1.0"
description = ""
authors = ["Orlovchik <svetl.orlovaa@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
pandas = "^2.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
